refactor(local_storage): use a local reference
this allows to not modify the map as much. We know that a given
reference will only ever be accessed by its thread (because it's the
only one to use that key). the map (and atomic ref) is only modified
when `set` is called on a key that is not used at all in this TLS
variable.